feat(trading): support fractional shares (quantity to 2 decimals)
Whole-share-only trading plus the 25% position cap was punishing; allow
buying/selling fractional shares (e.g. 0.95) so players can fill up to the
cap precisely.

- Quantity is now decimal end-to-end: Trade/Holding entities, buy/sell
  commands + events, MarketPriceInput, trading guard, read models/responses,
  and the holding-total aggregate.
- EF: trades.quantity + holdings.quantity -> numeric(18,2); migration
  SupportFractionalShareQuantities (int -> numeric, widening/safe).
- Validators reject more than 2 decimal places.
- Market volume stays a rounded long aggregate (summed as decimal first to
  avoid per-row truncation), so market read models/contracts are unchanged.
- Portfolio holdings sort moved client-side (single SELECT preserved; the
  SQLite query-count test provider can't ORDER BY decimal — Postgres can).

84 unit + 121 integration tests pass. Live: qty 0.95 accepted, 0.955 (3dp)
rejected with 400.

+++ src/Application/Features/Trading/BuyStock/BuyStockCommand.cs
@@ -3,5 +3,5 @@ using OsuStocks.Application.Common.Models;
 
 namespace OsuStocks.Application.Features.Trading.BuyStock;
 
-public sealed record BuyStockCommand(Guid UserId, Guid StockId, int Quantity)
+public sealed record BuyStockCommand(Guid UserId, Guid StockId, decimal Quantity)
     : IRequest<Result<BuyStockResponse>>;
