+++ tests/OsuStocks.Api.IntegrationTests/Infrastructure/InMemoryHoldingRepository.cs
@@ -38,7 +38,7 @@ internal sealed class InMemoryHoldingRepository : IHoldingRepository
         return Task.FromResult<IReadOnlyList<Holding>>(holdings);
     }
 
-    public Task<int> GetTotalQuantityByStockAsync(Guid stockId, CancellationToken cancellationToken = default)
+    public Task<decimal> GetTotalQuantityByStockAsync(Guid stockId, CancellationToken cancellationToken = default)
     {
         var total = _holdingsById.Values
             .Where(x => x.StockId == stockId)
